// src/generated/models/MoldingBuyoffStructureModel.ts
export interface ScrapLogged_x0028_Joe_x0029_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BatchCodeLegibleValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BatchCodeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Machine_x0023_Value {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Machine_x0020__x0023__x003a_AreaValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Restart_x002f_MoldChange_x003f_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface MaterialValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface MaterialTypeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface NTQValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface OData__x0023_ofCavsValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface MasterCardValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DefectDescriptionValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Mechanical_x0020_Defects_x003a_DValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface DimensionalDefectsValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Dimensional_x0020_Defects_x003a_Value {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface VisualDefectsValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Visual_x0020_Defects_x003a_DefecValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Deviation_x003f__x0028_Visual_x00Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DeviationPending_x003f__x0028_ViValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Deviation_x003f__x0028_DimensionValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DeviationPending_x003f__x0028_DiValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Deviation_x003f__x0028_MechanicaValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DeviationPending_x003f__x0028_MeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyOffAccepted_x003f_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyoffAccepted_x0028_DimensionalValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyoffAccepted_x0028_Visual_x002Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BatchCode_x0028_QC_x0029_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Shift_x0028_QC_x0029_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface PlantValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyoffAccepted_x0028_Mechanical_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ChangeoverComplete_x003f_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface MoldingBuyoffStructureBase {
  ID?: number;
  Title?: string;
  RunID?: string;
  DateCreated?: string;
  TimeIn?: string;
  WorkOrder_x0023_?: string;
  QtyScrapped?: string;
  "ScrapLogged_x0028_Joe_x0029_#Id"?: number;
  "BatchCodeLegible#Id"?: number;
  "BatchCode#Id"?: number;
  "Machine_x0023_#Id"?: number;
  "Machine_x0020__x0023__x003a_Area#Id"?: number;
  Part_x0023_?: string;
  PartDescription?: string;
  "Restart_x002f_MoldChange_x003f_#Id"?: number;
  "Material#Id"?: number;
  "MaterialType#Id"?: number;
  OtherMaterial?: string;
  "NTQ#Id"?: number;
  SprueWeight?: string;
  PartWeight?: string;
  "OData__x0023_ofCavs#Id"?: number;
  "MasterCard#Id"?: number;
  MoldingTech?: string;
  Comments?: string;
  "DefectDescription#Id"?: number[];
  "DefectDescription@odata.type"?: string;
  "Mechanical_x0020_Defects_x003a_D#Id"?: number[];
  "Mechanical_x0020_Defects_x003a_D@odata.type"?: string;
  "DimensionalDefects#Id"?: number[];
  "DimensionalDefects@odata.type"?: string;
  "Dimensional_x0020_Defects_x003a_#Id"?: number[];
  "Dimensional_x0020_Defects_x003a_@odata.type"?: string;
  "VisualDefects#Id"?: number[];
  "VisualDefects@odata.type"?: string;
  "Visual_x0020_Defects_x003a_Defec#Id"?: number[];
  "Visual_x0020_Defects_x003a_Defec@odata.type"?: string;
  "Deviation_x003f__x0028_Visual_x00#Id"?: number[];
  "Deviation_x003f__x0028_Visual_x00@odata.type"?: string;
  "DeviationPending_x003f__x0028_Vi#Id"?: number;
  DeviationNumber_x0028_Visual_x00?: string;
  DeviationDescription_x0028_Visua?: string;
  "Deviation_x003f__x0028_Dimension#Id"?: number[];
  "Deviation_x003f__x0028_Dimension@odata.type"?: string;
  "DeviationPending_x003f__x0028_Di#Id"?: number;
  DeviationNumber_x0028_Dimensiona?: string;
  DeviationDescription_x0028_Dimen?: string;
  "Deviation_x003f__x0028_Mechanica#Id"?: number[];
  "Deviation_x003f__x0028_Mechanica@odata.type"?: string;
  "DeviationPending_x003f__x0028_Me#Id"?: number;
  DeviationNumber_x0028_Mechanical?: string;
  DeviationDescription_x0028_Mecha?: string;
  "BuyOffAccepted_x003f_#Id"?: number;
  QCTech?: string;
  TimeOut?: string;
  "BuyoffAccepted_x0028_Dimensional#Id"?: number;
  QCTech_x0028_Dimensional_x0029_?: string;
  TimeOut_x0028_Dimensional_x0029_?: string;
  "BuyoffAccepted_x0028_Visual_x002#Id"?: number;
  QCTech_x0028_Visual_x0029_?: string;
  TimeOut_x0028_Visual_x0029_?: string;
  "BatchCode_x0028_QC_x0029_#Id"?: number;
  "Shift_x0028_QC_x0029_#Id"?: number;
  QCComments?: string;
  "Plant#Id"?: number;
  Last_x0020_Time_x0020_Out?: string;
  All_x0020_3_x0020__x0022_Yes_x00?: string;
  Attach?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  "BuyoffAccepted_x0028_Mechanical_#Id"?: number;
  QCTech_x0028_Mechanical_x0029_?: string;
  TimeOut_x0028_Mechanical_x0029_?: string;
  Date_x0020_Completed_x0020__x002?: string;
  Date_x0020_Completed_x0020__x0020?: string;
  Date_x0020_Completed_x0020__x0021?: string;
  FirstGoodPart?: string;
  "ChangeoverComplete_x003f_#Id"?: number;
  LastGoodPart_x0028_Ignore_x0029_?: string;
  ChangeoverTechnician?: string;
  Standardized_x0020_Part_x0020__x?: string;
  NextWorkOrder?: string;
  Modified?: string;
  Consolidated_x0020_Deviation_x00?: string;
  TimeInTest?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface MoldingBuyoffStructureRead extends MoldingBuyoffStructureBase {
  ScrapLogged_x0028_Joe_x0029_?: ScrapLogged_x0028_Joe_x0029_Value;
  BatchCodeLegible?: BatchCodeLegibleValue;
  BatchCode?: BatchCodeValue;
  Machine_x0023_?: Machine_x0023_Value;
  Machine_x0020__x0023__x003a_Area?: Machine_x0020__x0023__x003a_AreaValue;
  Restart_x002f_MoldChange_x003f_?: Restart_x002f_MoldChange_x003f_Value;
  Material?: MaterialValue;
  MaterialType?: MaterialTypeValue;
  NTQ?: NTQValue;
  OData__x0023_ofCavs?: OData__x0023_ofCavsValue;
  MasterCard?: MasterCardValue;
  DefectDescription?: DefectDescriptionValue;
  Mechanical_x0020_Defects_x003a_D?: Mechanical_x0020_Defects_x003a_DValue;
  DimensionalDefects?: DimensionalDefectsValue;
  Dimensional_x0020_Defects_x003a_?: Dimensional_x0020_Defects_x003a_Value;
  VisualDefects?: VisualDefectsValue;
  Visual_x0020_Defects_x003a_Defec?: Visual_x0020_Defects_x003a_DefecValue;
  Deviation_x003f__x0028_Visual_x00?: Deviation_x003f__x0028_Visual_x00Value;
  DeviationPending_x003f__x0028_Vi?: DeviationPending_x003f__x0028_ViValue;
  Deviation_x003f__x0028_Dimension?: Deviation_x003f__x0028_DimensionValue;
  DeviationPending_x003f__x0028_Di?: DeviationPending_x003f__x0028_DiValue;
  Deviation_x003f__x0028_Mechanica?: Deviation_x003f__x0028_MechanicaValue;
  DeviationPending_x003f__x0028_Me?: DeviationPending_x003f__x0028_MeValue;
  BuyOffAccepted_x003f_?: BuyOffAccepted_x003f_Value;
  BuyoffAccepted_x0028_Dimensional?: BuyoffAccepted_x0028_DimensionalValue;
  BuyoffAccepted_x0028_Visual_x002?: BuyoffAccepted_x0028_Visual_x002Value;
  BatchCode_x0028_QC_x0029_?: BatchCode_x0028_QC_x0029_Value;
  Shift_x0028_QC_x0029_?: Shift_x0028_QC_x0029_Value;
  Plant?: PlantValue;
  BuyoffAccepted_x0028_Mechanical_?: BuyoffAccepted_x0028_Mechanical_Value;
  ChangeoverComplete_x003f_?: ChangeoverComplete_x003f_Value;
  Author?: AuthorValue;
  Editor?: EditorValue;
}